requirements.txt
streamlit
google-generativeai
PyMuPDF
reportlab
json5
